requirements.txt
pandas
numpy
joblib
matplotlibgit